[Rename]: friday (2).py to Friday.py
